Probability of levee failure
Calculated the probability of failure per leveed area from the RAND model. The flat file located in data\randProbFailure folder.

The calculation uses a stripped-down version of the dm0 script by RAND. Exploration related to how RAND calculated this metric is in data-raw\randModel\nceas\exploreLeveeProbFailure.R

Likely will need to identify which leveed area we share with the RAND model--only those areas will have a probability calculation as we are not calculating it ourselves.